Update home with new layout
